Implemented Abilities (not tested)
